// types/CuttingHistoryType.ts
export type CuttingHistoryType = {
  id: string;
  staff: string;
  processNumber: string;
  cuttingDate: string;
  itemName: string;
  itemType: string;
  client: string;
  totalQuantity: number;
  comment: string;
  category: string;
  productId: string;
  quantity: number
  serialNumber: number;
};